// apps/backend/src/types/enums.ts
export enum AccountType {
  BANK = 'bank',
  CRYPTO = 'crypto',
  CREDIT_CARD = 'credit_card'
}

export enum Currency {
  EUR = 'EUR',
  USDC = 'USDC',
  ARS = 'ARS',
  USD = 'USD'
}

export enum TransactionType {
  INCOME = 'income',
  TRANSFER = 'transfer',
  EXPENSE = 'expense',
  CREDIT_CARD_PAYMENT = 'credit_card_payment'
}